Moving MonoCustomAttrValue types to custom-attrs-internals.h
